Tests/LibWeb/Text/expected/css/abspos-static-position-after-inline-start-edges.txt

direct: abspos block offset in wrapper: 20
restashed: abspos block offset in wrapper: 20
control: abspos block offset in wrapper: 0
